fix vec3d setting parsing

+++ src/main/java/meteordevelopment/meteorclient/settings/Vector3dSetting.java
@@ -44,7 +44,7 @@ public class Vector3dSetting extends Setting<Vector3d> {
     protected Vector3d parseImpl(String str) {
         try {
             String[] strs = str.split(" ");
-            return new Vector3d(Integer.parseInt(strs[0]), Integer.parseInt(strs[1]), Integer.parseInt(strs[2]));
+            return new Vector3d(Double.parseDouble(strs[0]), Double.parseDouble(strs[1]), Double.parseDouble(strs[2]));
         } catch (IndexOutOfBoundsException | NumberFormatException ignored) {
             return null;
         }
